refactor: update adapters package to export only wagmi temporarily
