+++ packages/adapters/index.ts
@@ -1,0 +1,4 @@
+import '@web3modal/polyfills'
+
+export { EVMWagmiClient } from './src/client.js'
+export type { CoreConfig, ReactConfig, Web3ModalClientOptions } from './src/client.js'
