+++ packages/adapters/src/connectors/AuthConnector.ts
