+++ packages/adapters/src/connectors/EmailConnector.ts
