+++ packages/adapters/src/utils/constants.ts
