+++ packages/adapters/src/utils/defaultWagmiCoreConfig.ts
