+++ packages/adapters/src/utils/defaultWagmiReactConfig.ts
