+++ packages/adapters/tsconfig.wagmi.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "baseUrl": ".",
+    "outDir": "./dist/evm/wagmi/esm",
+    "declarationDir": "./dist/evm/wagmi/types",
+    "declaration": true,
+    "declarationMap": true
+  },
+  "extends": "../../tsconfig.json",
+  "include": ["wagmi.ts", "evm/wagmi/**/*.ts"]
+}
